chore: refactoring as per comments
